lesson-2: added blank NoteEditFragment.kt, refactor packages
