+++ app/src/main/java/com/rogoz208/notesv2/ui/screens/notes/edit/NoteEditFragment.kt
@@ -1,0 +1,16 @@
+package com.rogoz208.notesv2.ui.screens.notes.edit
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.View
+import by.kirich1409.viewbindingdelegate.viewBinding
+import com.rogoz208.notesv2.databinding.FragmentNoteEditBinding
+
+class NoteEditFragment : Fragment() {
+    private val binding by viewBinding(FragmentNoteEditBinding::bind)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+}
