+++ app/src/main/java/com/rogoz208/notesv2/ui/screens/notes/list/OnItemClickListener.kt
@@ -1,4 +1,4 @@
-package com.rogoz208.notesv2.ui.screens.notes
+package com.rogoz208.notesv2.ui.screens.notes.list
 
 import android.view.View
 import com.rogoz208.notesv2.domain.entities.NoteEntity
