+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,6 @@
     <string name="notes_bottom_nav">Notes</string>
     <string name="reminders_bottom_nav">Reminders</string>
     <string name="settings_bottom_nav">Settings</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
